django/BankAccount/requirements.txt
cryptography==2.3
psycopg2-binary==2.8.4
django==3.2
django-crum==0.7.5
django-encrypted-model-fields==0.5.8
djangorestframework==3.11.0
PyJWT==1.7.1
djangorestframework-simplejwt==4.4.0
django-cookies-samesite==0.5.1
Pillow==7.0.0
gunicorn==20.0.4
whitenoise==5.0.1